+++ index.js
@@ -1,5 +1,6 @@
 module.exports = {
   IPCalculator: require('./lib/IPCalculator'),
   ConnTester: require('./lib/ConnTester'),
-  IFaceConfigurator: require('./lib/IFaceConfigurator')
+  IFaceConfigurator: require('./lib/IFaceConfigurator'),
+  IPRouter: require('./lib/IPRouter')
 }