+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-# v1.3.0
+# v1.5.0
+- Added methods getIpForward and setIpForward to IPRouter.
+- Some bug fixes in IPRouter module.
+- IPRouter exposed in main module.
+# v1.4.0
 - Added option to choose table in getRoutes, in options object.
-# v1.2.0
+# v1.4.0
 - Exposed method convertNetmaskToCidr from *IPCalculator*.
 - Exposed method convertCidrToNetmask from *IPCalculator*.
 - Method addAddress implemented for IFaceConfigurator. Now the module can add IP address to an interface..